refactor(indexer): migrate SpIndexerV2 to its own module
Move the SpIndexerV2 implementation from lib.rs into a dedicated v2
module to improve code organization and maintainability. All existing
functionality is preserved while reducing the size of lib.rs.

This commit restructures the codebase by moving the SpIndexerV2
implementation and its related code (including indexes.rs) into a
dedicated v2 module. The code remains functionally identical, but is now
better organized with clearer module boundaries. All import paths in
dependent files (oracles/tweaks/blindbit.rs, wallet/lib.rs,
wallet/signers.rs) have been updated to reference the new module
location.
